addrxlat: Convert read_caps to a callback
It was OK to store the capabilities value directly in the callback
structure when there was only one. But if callbacks are chained,
then it's not clear how the combined capabilities of the complete
chain of get-page callbacks should be derived from individual
values. An implementation may want to:

  - simply use the next callback, or
  - enhance the capabilities, or
  - override the capabilities

to match what is done in the get-page callback. A callback can
easily do all of that.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>