Change file set attribute layout
The original layout was like this:

- file
  - fdset
    - number
    - 0
    - 1
    - ...
    - $number_1

The numbered attributes contained the corresponding file
descriptor (as a number).

The problem with that layout was that there was no logical place
for additional per-file metadata, such as the PFN range for each
file or other format-specific fields from the file header.

In the new layout, the numbered attributes are directories, each
with a `fd` sub-attribute:

- file
  - set
    - number
    - 0
      - fd
    - 1
      - fd
    - ...
    - $number_1
      - fd

Further per-file attributes can be added next to the `fd`
sub-attribute.

Besides, a similar layout is already used for attributes in the
`cpu` sub-tree.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>