Fix search for a clear bit in a PFN bitmap
The bitwise NOT operator (~) extends the operand to an int, so the
high bits are always set after applying it to an unsigned char.
As a result, the scan never skips the first byte if the pointer is
unaligned.

The resulting region array is not incorrect, but much larger than
necessary.

Fixes: d68e84464410469b228df58e9650be5cfe64fee4
Signed-off-by: Petr Tesarik <ptesarik@suse.com>