Clean up and micro-optimize bitmap scan functions
It's easier to understand a check against 0xff than typecasts and
negations. This can be done in skip_set_msb0(), where the value in
the conditional is otherwise unused.

For skip_set_lsb0(), adjust the argument to ctz() instead, so that
the value computed for the conditional can be reused.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>