elfdump: Fix multiple corner cases in segment search
First, skip zero-sized segments. An address is not really
contained in a zero-sized segment.

Second, start address + size may overflow, so let's reorder the
calculations so that we cope with that case. Note that one check
uses size minus one, but that's OK now, because zero size was
already handled.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>