elfdump: Honour file.zero_excluded
There is a difference between a partial page and a page that is
completely missing from the ELF dump. We are currently filling
missing parts with zeroes, regardless of the file.zero_excluded
attribute. This is OK according to the ELF specification, but
makedumpfile reduces filesz even if the filtered out page is not
all zeroes, so these are indeed excluded pages and should be
reported as such unless file.zero_excluded is set.

The elf-multiread test relied on the broken behaviour, and had to
be fixed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>