tests: Add missing checks for large regions on AArch64
A level 1 table with 16k pages can specify a BLOCK type, which
then defines a 64G region. This should be tested.

A level 1 table with 64k pages can also specify a BLOCK type,
which then defines a 4T region. This should also be tested.

Besides, I have fixed an incorrect comment for the 32M region with
16k pages.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>