AArch64: Fix handling of block descriptors
Fix the check for minimum table level where block descriptors are
allowed. They were incorrectly rejected in level zero tables with
a 4 KiB granule, because the page size is 12 bits, not 9 bits.

However, the check can be simplified as described in the comment
to MAX_REGION_MASK.

Note that 5-level paging has not been tested for AArch64, and the
test for invalid translations verifies the incorrect behaviour.
Another translation level must be added to fix it.

Fixes: dd1490fb68f9b8fca6dfc876aed95c64cc24a3b0
Signed-off-by: Petr Tesarik <ptesarik@suse.com>