AArch64: Add formats for LPA and LPA2 descriptors
The handling of AArch64 translation tables is still broken,
because it can handle 52-bit input addresses and allows large
regions that depend on the LPA/LPA2 feature, but it does not
handle 52-bit output addresses, which depend on the same hardware
feature.

As it happens, the issue is more complicated, because the
descriptor formats were changed to make room for the extra output
address bits. Instead of putting the complexity to pgt_aarch64(),
let's admit that these page table formats are different and expose
them in the API.

As a bonus, page table entries with non-zero should-be-zero bits
and/or non-zero Shareability bits can still be handled correctly
if the original PTE format is used.

There is no single answer to the Great Question of Life, the
Universe, and Everything...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>