addrxlat: Move the code of addrxlat_pteval_shift() to the library
This utility function is defined as an inline function. This is
sub-optimal, because programs must be recompiled after adding a
new PTE format.

However, the function is called from first_step_pgt_generic(),
which is in the hot path. To keep original performance, make a
private inline function that should be used within the library and
to implement the public function.

Note: Update libtool versioning, because this is the first
addition to libaddrxlat ABI since release of 0.5.0.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>