python/kdumpfile: add from_pointer method to clone native context
When using libdumpfile in a project that uses the library directly but
also has Python extensions that need access to it, we want to clone the context
instead of opening a new one so that we share resources and don't need
to open the dump file multiple times.

This is an interface that does allow the developer to shoot themselves
in the foot if an invalid pointer is passed, but this is a common design
pattern.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>