Ignore empty sysname, release, version or machine in utsname
The domain name and nodename may be empty in uname, but other
fields should be treated as undefined if they are empty.

In particular, the vmconvert utility on s390x sets only the
machine field in the output file, leaving other fields empty,
which makes the rest of the library believe that the kernel
version is in fact known, failing later when the version_code
cannot be parsed. This error message is observed:

Cannot get linux.version_code: Invalid kernel version:

(The empty space after the last colon is the empty string that
should be a valid kernel version.)

Signed-off-by: Petr Tesarik <ptesarik@suse.com>