Fix check for empty probed partition in cache_discard()
Discarded cache entries are added to the unused partition of the
cache. The beginning of this partition is equal to the end of the
probe partition (eprobe). If the probe partition is empty, its
beginning is also equal to the cache split point. Since we're
inserting a new entry after eprobe, the split point must be
updated.

However, the condition is incorrect, because eprobe can also reach
the split point if the probe partition takes up the whole cache,
because the list is circular. Moving the split point is fatal in
that case, because it shifts cache partitions by one, and the
unused partition then contains an entry with NULL data, which
fails later when the cache tries to use it.

Reported-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Petr Tesarik <petr@tesarici.cz>