cache: Silence a false positive compiler warning
GCC complains that cs->eprobe may be uninitialized in
get_missed_entry(). The program logic makes sure that it is always
initialized if it is used, but that's not known to the compiler.
Instead of initializing the field to a bogus value, I have
reshuffled the code to make it clear when the field is actually
used.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>