Cache: Reuse data from the unused partition if available
If reinit_entry() is called while there are no unused cached
entries in either precious or probed partition, it will fail.
This can happen if all entries with a zero reference count are in
the unused partition of the cache.

If there are less entries in the precious partition, probe
partition and in flight than the cache capacity, then more than
half of all entries must be in the unused partition. But then we
must be either still consuming first half of the initial batch, or
an entry must have been discarded. In both cases, the unused entry
contains a valid data pointer.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>