Fix hashing of unaligned strings
The hash routines try to be fast by hashing unsigned longs.
However, this breaks on architectures that do not allow unsigned
access (e.g. SPARC64). Using a typecast to a packed structure
tells the compiler to emit code for unaligned access where
necessary. That way, there is no performance hit to architectures
that implement unaligned access in hardware.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>