diskdump: Support flattened diskdump format
Allow reading flattened diskdump files directly.

The mapping from rearranged file offsets to flattened file offsets is
implemented using libaddrxlat's addrxlat_map_t, which is normally used to
map an address to a translation method. Here it maps the rearranged file
offset to an index into an array of flattened file offset. This is a hack,
but it allows to reuse the non-trivial logic in addrxlat_map_set(), which
avoids the need to walk all segments in case some (rearranged) offset was
written more than once.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>