diskdump: Directly map buffers from flattened file cache if possible
Check if a flattened segment completely covers a requested chunk and if
yes, map the underlying file cache buffer instead of copying it to a
dynamically allocated temporary buffer.

Since makedumpfile writes a flattened segment only after adding a complete
compressed page to its output buffer, this is the normal case. This change
saves a lot of allocating, copying and freeing.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>