devmem: Clear error message after ignoring vmcoreinfo read errors
Do not keep error messages if get_vmcoreinfo() fails. Reading VMCOREINFO
fail for a number of reasons, but none of them is fatal. If initialization
fails later on because of insufficient information, the VMCOREINFO error
message would be useful. However, if it fails for an unrelated reason,
this same error message is highly confusing.

Just clear it.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>