ELF: When opening /proc/kcore, use VMCOREINFO from running system
If necessary, read VMCOREINFO from the currently running system, using data
from /sys/kernel/vmcoreinfo.

The check for /proc/kcore is a bit hackish, because libkdumpfile generally
knows only the file descriptor. It is assumed that a dump is /proc/kcore if:

- it is an ELF dump,
- it is located on the proc filesystem,
- it contains NT_TASKSTRUCT.

The last condition is meant to distinguish between /proc/kcore (in a normal
kernel) and /proc/vmcore (in a crash kernel).

Note that VMCOREINFO is included in /proc/kcore since Linux v2.6.24-rc1
(commit fd59d231f81cb02870b9cf15f456a897f3669b4e), so this logic is only a
fall back for ancient kernels.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>