cache: Optimize data reclaim for missed entries
While wondering if a variant of the NULL-data pointer bug is possible on a
cache miss without a ghost entry, it turns out that data can never be
reclaimed from an entry in the unused partition. If there is such an entry,
it is the one that was already chosen as the return value.

If entry->data is NULL in get_missed_entry(), then another cache entry must
be evicted, and the condition in reclaim_data() will never be met. To save
people from wondering if there is a bug, call evict_entry() directly.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>