tests: skip tests which apply to disabled compression
If configured --without-libzstd, for example, the diskdump-basic-zstd
test will return an ERROR code, causing "make check" to fail. Even using
XFAIL_TESTS will not resolve the error, because the return code is
ERROR, not FAIL.

Instead, conditionally include the tests based on whether we are
compiling with each compression format. This way, we don't test
unsupported features.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>