Fix file cache test for 32-bit architectures
If 64-bit file offsets are selected with _FILE_OFFSET_BITS on a 32-bit
architecture, the default mmap() call takes a 64-bit off_t, but dlsym()
returns a pointer to a function that takes a 32-bit off_t.

To fix it:

- always call original mmap64() if it is available,
- use XSTRINGIFY(mmap) instead of "mmap".

The latter is needed, because some systems define mmap as a macro which
expands to another identifier.

Fixes: #80
Signed-off-by: Petr Tesarik <petr@tesarici.cz>