diskdump: Always allocate a flattened offset map
Prepare to support a mix of flattened and rearranged files in a file set by
checking individual files in the set with flatmap_isflattened() instead of
the flatmap pointer.

This change also simplifies initialization, but the disk dump private data
is now always allocated (and freed), even if the dump file turns out to be
a different format. That should be a very minor regression, though.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>