flatmap: Allow to use the flatmap API with non-flattened files
The flatmap API now gets enough information to know if a file is flattened
and (where appropriate) read directly from a non-flattened file. In fact,
this is what elfdump_pread() and elfdump_get_chunk() already do.

Moreover, the check adds only little overhead (especially if the function
is inline), so rename the original functions (which can read only flattened
files) by adding a _flat suffix to them and make the API usable with
non-flattened files.

Remove the format-specific diskdump and elfdump functions, which do
essentially the same job as flatmap_pread() and flatmap_get_chunk()
after this change.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>