flatmap: Initialize flatmap offsets only once
The flattened offset mappings is initialized when probing for a format that
can be flattened. The downside is that for a flattened diskdump, the
initialization is done twice: first when probing for ELF dump (just to find
out it is not an ELF dump, so the mappings are immediately discarded) and
then again when probing for diskdump.

To fix it, move the flatmap object to struct kdump_shared and initialize it
before probing for format.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>