tests/mkdiskdump: Fix VMCOREINFO, ELF notes and eraseinfo
Store the VMCOREINFO, ELF notes and eraseinfo contents in the resulting
dump file.

Note that all of this data is currently stored in the sub-header.
Makedumpfile also uses the sub-header for VMCOREINFO and ELF notes, but it
saves eraseinfo at the end of the file, so the generated files are valid
but organized slightly differently than real-world samples.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>