x86_64: Fix ktext mapping if it starts above 0xffffffffa0000000
When all of kernel text is relocated beyond the pre-kASLR highest virtual
address, the scan for region end is not even attempted.

If the kernel text start address is found above LINUX_KTEXT_END_NOKASLR,
unconditionally scan for the region end up to LINUX_KTEXT_END.

The corresponding test case is derived from a real-world dump.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>